Add kinds
