+++ src/EditorFeatures/Core/Extensibility/NavigationBar/NavigationBarItems/RoslynNavigationBarItem.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#nullable disable
-
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis.Text;
 
@@ -12,7 +10,7 @@ namespace Microsoft.CodeAnalysis.Editor.Extensibility.NavigationBar
     /// <summary>
     /// Base type of all C#/VB navigation bar items.  Only for use internally to roslyn.
     /// </summary>
-    internal abstract class RoslynNavigationBarItem : NavigationBarItem
+    internal abstract partial class RoslynNavigationBarItem : NavigationBarItem
     {
         public readonly RoslynNavigationBarItemKind Kind;
 
@@ -21,7 +19,7 @@ namespace Microsoft.CodeAnalysis.Editor.Extensibility.NavigationBar
             string text,
             Glyph glyph,
             IList<TextSpan> spans,
-            IList<NavigationBarItem> childItems = null,
+            IList<NavigationBarItem>? childItems = null,
             int indent = 0,
             bool bolded = false,
             bool grayed = false)
